Move StateAccess to API
