+++ inventory-framework-api/src/main/java/me/devnatan/inventoryframework/StateAccess.java
